requirements.txt
# filepath: /Users/oskar/Project/home-control-panel-eink/requirements.txt
pillow==11.2.1
# MQTT client for dynamic updates
paho-mqtt==1.6.1
# Weather API uses only standard library (urllib)
